Parameratizes classes to use hiera / console lookups
Now assumes the module is installed along with other PE modules in
/opt/puppet/share/puppet/modules, which makes the rest of the
determining factors easier for created the pe_bootstrap environment on
the master.

It also now (in theory) has some code to do, by default, ensure agents
are running the same version of PE as the master compiling the catalog,
those settings can be overwritten via hiera or console. The same
settings are in place now to allow for the puppet mastter the agent
should configure / build against (allow for Master A to point agents to
Master B, useful for migrations).